Extract command handlers
